refactor topology into folder (getting too long)
